Make product images clickable
Made images on product cards clickable like the button below them. When I accidentally go to discordlabs.org instead of bots.discordlabs.org it takes like 0.2s to scroll down the page because the images are huge and it wastes my time. Images should be clickable, so I wrote the code for you. You're welcome.